20230711_00:15 | hope mariadb is done:(

+++ srcs/requirements/mariadb/tools/script.sh
@@ -1,23 +1,30 @@
 #!/bin/bash
 
-# if [ ! -f "/setup.sql" ]; then
-/usr/bin/mysqld_safe --datadir=/var/lib/mysql &
+/usr/bin/mysqld_safe --datadir=/var/lib/mysql --skip-networking --nowatch
 
-while ! mysqladmin ping --silent; do
+counter=0
+while ! mysqladmin ping --silent && [[ $counter -lt 30 ]]; do
 	sleep 1
+	((counter++))
+	echo "Waiting for mysql server to start. $counter/30 seconds passed."
 done
+
+if [ ! -f "/root_pw.sql" ]; then
 
 touch root_pw.sql
 
 echo "UPDATE mysql.user SET Password=PASSWORD('$MYSQL_ADMIN_PW') WHERE User='root' ;" >> root_pw.sql
 echo "DELETE FROM mysql.user WHERE User='' ;" >> root_pw.sql
-echo "DELETE FROM mysql.user WHERE User=$MYSQL_ADMIN_USER' AND Host NOT IN ('localhost', '127.0.0.1', '::1') ;" >> root_pw.sql
+echo "DELETE FROM mysql.user WHERE User='$MYSQL_ADMIN_USER' AND Host NOT IN ('localhost', '127.0.0.1', '::1') ;" >> root_pw.sql
 echo "DROP DATABASE IF EXISTS test ;" >> root_pw.sql
-echo "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%' ;" >> root_pw.sql
+echo "DELETE FROM mysql.db WHERE Db='test' ;" >> root_pw.sql
 echo "FLUSH PRIVILEGES ;" >> root_pw.sql
 
 mysql -u root < root_pw.sql
 
+fi
+
+if [ ! -f "/setup.sql" ]; then
 
 touch setup.sql
 
@@ -26,13 +33,15 @@ echo "CREATE USER IF NOT EXISTS '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PW' ;" >
 echo "GRANT ALL PRIVILEGES ON $MYSQL_DB_NAME.* TO '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PW' WITH GRANT OPTION ;" >> setup.sql
 echo "FLUSH PRIVILEGES ;" >> setup.sql
 
+# mysql -u $MYSQL_ADMIN_USER -p$MYSQL_ADMIN_PW < setup.sql
+mysql < setup.sql
 
-mysql -u $MYSQL_ADMIN_USER -p$MYSQL_ADMIN_PW < setup.sql
+fi
 
-while ! mysqladmin ping --silent; do
-	sleep 1
-done
-#fi
+# while ! mysqladmin ping --silent; do
+# 	sleep 1
+# done
+# fi
 
 mysqld
